% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blackjack.R
\name{blackjack}
\alias{blackjack}
\title{blackjack}
\usage{
blackjack()
}
\value{
Fun
}
\description{
A blackjack game with the most important features
including betting, ace conversion and blackjack
}
\examples{
\dontrun{
blackjack()
}
}
