% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/magicRball.R
\name{magicRball}
\alias{magicRball}
\title{magicRball}
\usage{
magicRball()
}
\value{
An answer to your question
}
\description{
If it is wisdom you seek, this function will illuminate your path.
}
\examples{
\dontrun{
magicRball()
}
}
