% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slotmachine.R
\name{slotmachine}
\alias{slotmachine}
\title{slotmachine}
\usage{
slotmachine()
}
\value{
Called for fun. No return value.
}
\description{
Take a spin and gamble all your points. This slot machine has certainly not
been rigged, you can check the source code. The starting balance
is 100 points and can be increased up to five fold per spin.
Run the function without input to play.
}
\examples{
\dontrun{
slotmachine()
}
}
