% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tictactoe.R
\name{tictactoe}
\alias{tictactoe}
\title{Tic Tac Toe}
\usage{
tictactoe()
}
\value{
Called for fun. No return value.
}
\description{
A multiplayer game of Tic Tac Toe. The function automatically alternates
between turns and declares a winner when a win condition is met.
Run the function without input and follow the steps in the console,
alternating between players.
}
\examples{
\dontrun{
tictactoe()
}
}
