% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blackjack.R
\name{blackjack}
\alias{blackjack}
\title{blackjack}
\usage{
blackjack()
}
\value{
Called for fun. No return value.
}
\description{
A blackjack game with the most important features
including betting, ace conversion and blackjack.
Run the function without any inputs and follow the steps in the console.
}
\examples{
\dontrun{
blackjack()
}
}
