% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/magicRball.R
\name{magicRball}
\alias{magicRball}
\title{magicRball}
\usage{
magicRball()
}
\value{
Called for fun. No return value.
}
\description{
This magic R ball will provides wisdom to anyone who seeks it.
It has numerous data science inspired responses to any question.
Run the function without any inputs and follow the steps in the console.
}
\examples{
\dontrun{
magicRball()
}
}
